### 2019-11-16 #1894 @whedon
[PRE REVIEW]: Dryft: A Python and MATLAB package to correct drifting ground reaction force signals during treadmill running
**Submitting author:** @alcantarar (<a href="http://orcid.org/0000-0002-8539-711X">Ryan Alcantara</a>)
**Repository:** <a href="https://github.com/alcantarar/dryft" target ="_blank">https://github.com/alcantarar/dryft</a>
**Version:** v1.0.0
