### 2019-11-16 b2a80dd
add dois for joss pub
