+++ documentation/JOSS_submission/paper.bib
@@ -6,6 +6,8 @@
   number={2},
   pages={764--769},
   year={1998},
+  doi = {10.1152/jappl.1998.85.2.764},
+  url = "https://doi.org/10.1152/jappl.1998.85.2.764"
 }
 
 @article{fukuchi2017public,
@@ -15,7 +17,9 @@
   volume={5},
   pages={e3298},
   year={2017},
-  publisher={PeerJ Inc.}
+  publisher={PeerJ Inc.},
+  doi = {10.7717/peerj.3298},
+  url = "https://doi.org/10.7717/peerj.3298"
 }
 
 @article{paolini2007testing,
@@ -26,7 +30,9 @@
   number={3},
   pages={404--411},
   year={2007},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi = {10.1016/j.medengphy.2006.04.002},
+  url = "https://doi.org/10.1016/j.medengphy.2006.04.002"
 }
 
 @article{sloot2015comprehensive,
@@ -37,7 +43,9 @@
   number={6},
   pages={610--616},
   year={2015},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi = {10.1016/j.medengphy.2015.03.018},
+  url = "https://doi.org/10.1016/j.medengphy.2015.03.018"
 }
 
 @article{riley2008kinematics,
@@ -48,7 +56,9 @@
   number={6},
   pages={1093--1100},
   year={2008},
-  publisher={LWW}
+  publisher={LWW},
+  doi = {10.1249/MSS.0b013e3181677530},
+  url = {https://www.ncbi.nlm.nih.gov/pubmed/18460996}
 }
 
 @Manual{scipy:detrend,
