### 2019-11-16 c20edef
remove broken elsevier links, replaced with pubmed
