+++ documentation/JOSS_submission/paper.bib
@@ -32,7 +32,7 @@
   year={2007},
   publisher={Elsevier},
   doi = {10.1016/j.medengphy.2006.04.002},
-  url = "https://doi.org/10.1016/j.medengphy.2006.04.002"
+  url = "https://www.ncbi.nlm.nih.gov/pubmed/16759895"
 }
 
 @article{sloot2015comprehensive,
@@ -45,7 +45,7 @@
   year={2015},
   publisher={Elsevier},
   doi = {10.1016/j.medengphy.2015.03.018},
-  url = "https://doi.org/10.1016/j.medengphy.2015.03.018"
+  url = "https://www.ncbi.nlm.nih.gov/pubmed/25921721"
 }
 
 @article{riley2008kinematics,
