### 2019-11-16 #1894 @alcantarar
:wave: @labarba , I tested the whedon paper preview and it compiled, do you have info on what this error is? I've also added DOIs for the references in the latest commit.
