### 2019-11-16 #1894 @labarba
@Kevin-Mattheus-Moerman -- Could you handle this submission as editor?
**Dryft: A Python and MATLAB package to correct drifting ground reaction force signals during treadmill running**
https://github.com/alcantarar/dryft
