### 2019-11-17 #1894 @arfon
> @openjournals/dev -- we're experiencing some compilation issues. Help?

I accidentally broke Whedon last night. Fixed now.
