### 2019-11-17 #1894 @agricolab
Would be interested and available for review.
