### 2019-11-20 #1894 @Kevin-Mattheus-Moerman
@modenaxe @demotu @jimmyDunne @anmuller @Ipuch @cpontonn would you be interested in reviewing this work for JOSS?
