### 2019-11-24 9eb26a0
add filter cutoff correction, fixes #22. see issue for more detials
