### 2019-11-27 #1910 @agricolab
Authors submitted a well-written manuscript for a well-documented toolbox on drift correction for oscillations generated by ground reaction force measurements. There are only four aspects that imho require action, the rest of my comments can be considered mere recommendations. These aspects are

- [ ] [discuss performance claims](#pf)
- [ ] [add tests](#at)
- [ ] [pin panda dependency](#pd)
- [ ] [revive dead link](#dl)

## Matlab

### Author might consider to add more documentation on installation for Matlab

Currently, the author redirects to the general Matlab documentatio for adding packages, but sth.along
```
git clone https://github.com/alcantarar/dryft.git
cd dryft
echo "addpath(\"$(pwd)\"); savepath(); exit;" > import_dryft.m
matlab -r 'run("import_dryft.m")'
```
might be helpful nonetheless, altough i am aware it'd probably would not to be refined for different platforms. If installation is done in this permanent fashion, using namespaces is  sensible, see next comment. An alternative is omitting the `savepath(); exit;` and make the user run `import_dryft.m` everytime they want to use dryft.

### Author might consider to use namespaces for Matlab

This can be achieved by putting all functions into a folder e.g. `~/MATLAB/+dryft`, which allows after `addpath("~/MATLAB")` to call matlab functions using dot.notation, e.g. `dryft.detrend`. See also https://de.mathworks.com/help/matlab/matlab_oop/scoping-classes-with-packages.html#brf3g8k This is a recommendation, only because namespaces are barely used in the Matlab ecosystem. It is good software practice, nonetheless.

## Python

### Author should pin the panda dependency as `pandas >= 0.24` <a name="pd"></a>

The package uses pandas Datafram.to_numpy which was [introduced in 0.24.0](https://pandas.pydata.org/pandas-docs/stable/reference/api/pandas.DataFrame.to_numpy.html)

### Author might consider to split installation recommendations into "classical" and "with environment"

I appreciate that an environment.yaml was supported, but am uncertain why the author suggests a conda environment as recommended installation. Are there any dependencies that require conda? Otherwise, conda would become a hidden dependency, yet would actually not be required at all. All requirements install fine from pip, and the environment can be set-up with virtualenv or similar, too. I would argue, it should be left to the user how they want to set up their environment, if at all.

### Author might consider to compact snippets and add instructions for downloading repo

Everyone benefits from compact and complete copy-pastable bash snippets, and i recommend sth along

```{bash}
git clone https://github.com/alcantarar/dryft.git
cd dryft
conda env create -f environment.yml
conda activate dryft-env
python setup.py install
```

## General

### Author should discuss performance <a name="pf"></a>

Author seems to suggest his approach is superior to other methods, as

> dryft does not assume a constant drift for a given step, instead spline interpolating between aerial phases. The dryft package differs from currently available signal correction methods, which can only account for linear drift or a constant offset.

While i would agree with the author that spline interpolation is theoretically better, author might want to add some numbers to this claim.

### Author should revive dead link <a name="dl"></a>

The link in https://github.com/alcantarar/dryft/tree/master/documentation/API_Documention_by_pdoc3 to the Pdoc3 page is apparently dead

### Author should provide automated tests <a name="at"></a>

This reviewer failed to find any automated tests or instructions for manual testing for Matlab or Python.

### Author is thanked for having created sufficient documentation

pdoc3 builds the html well
the docstrings are well written and concise, and sufficient contain examples

### Author might want to streamline example and add tests

Author includes an example "of how to use the software" in the readme. Author might consider to move the sample.py and the data within the package. For python, e.g. into the folder `dryft/example`. This would allow a snippet along the lines of `python -m dryft.example.sample` to run the example, and could also constitute a basic smoke test. The datafile can be found e.g using

```
from pathlib import Path
# Read in data from force plate
GRF = pd.read_csv(Path(__file__).parent /
                  'custom_drift_S001runT25.csv', header=None)
```

### Author might want to use full names in acknowledgments

The authors acknowledges various people by acronym only. Unless this is to maintain anonymity, it counters the goal of acknowledging someone, and author might want to use full names.
