### 2019-11-27 131a046
fixes #31 dead link in pdoc3 readme
