+++ documentation/API_Documention_by_pdoc3/README.md
@@ -5,4 +5,4 @@
 3. Navigate to repository locally: `cd path/to/dryft`
 4. Use pdoc3: `python -m pdoc --html dryft`
 
-Html files should be located in a new folder called 'html'. Pdoc3 page is [here](pdoc3.github.io/pdoc)
+Html files should be located in a new folder called 'html'. Pdoc3 page is [here](https://pdoc3.github.io/pdoc)
