### 2019-11-27 160ad46
fixes #33, acknowledgements comments in https://github.com/openjournals/joss-reviews/issues/1910
