### 2019-11-27 cb4ec64
fixes #28, require pandas>=0.24.0 in setup.py/environment.yml per https://github.com/openjournals/joss-reviews/issues/1910
