### 2019-11-27 cd51d65
fixes #26, fixes #27. reorg folders to add matlab namespace, elaborate on addpath in matlab readme, per https://github.com/openjournals/joss-reviews/issues/1910
