### 2019-11-27 0d0d7a4
clarify matlab readme as part of #26
