### 2019-11-27 9c6a551
make conda env install/activation more concise in readme per #30 and https://github.com/openjournals/joss-reviews/issues/1910
