+++ README.md
@@ -33,18 +33,13 @@ The package requires the following dependencies:
 
 I recommend using the [Anaconda](https://www.anaconda.com/distribution/#download-section) to setup a Python 3.6.7 
 environment to use this package. If you wish to setup a new environment, an Anaconda [environment](environment.yml) 
-file is included to automatically install most dependencies. To setup a new environment for `dryft`, type the following 
-command into Anaconda Prompt:
+file is included to automatically install most dependencies. 
+You can create/activate an anaconda environment and download dependencies: 
 ```
+git clone https://github.com/alcantarar/dryft.git
+cd dryft
 conda env create -f path\to\environment.yml
-```
-This will create a new environment named `dryft-env`, which can be actived by typing:
-```
 conda activate dryft-env
-```
-Once the environment is setup and activated, `dryft` can be installed by cloning/downloading
-this repository and then running:
-```
 python setup.py install
 ```
 
