### 2019-11-27 d362ea0
clarify readme instructions for anacona, fixes #30
