### 2019-11-27 002d4cd
rename sample.py and move into package as test.py, update readme, fixes #35, fixes #32
