+++ dryft.egg-info/requires.txt
@@ -1,0 +1,4 @@
+numpy
+pandas>=0.24.0
+matplotlib
+scipy
