+++ dryft.egg-info/SOURCES.txt
@@ -1,0 +1,10 @@
+README.md
+setup.py
+dryft/__init__.py
+dryft/plot.py
+dryft/signal.py
+dryft.egg-info/PKG-INFO
+dryft.egg-info/SOURCES.txt
+dryft.egg-info/dependency_links.txt
+dryft.egg-info/requires.txt
+dryft.egg-info/top_level.txt

[…]
