### 2019-11-27 9603b4b
remove some build files
