+++ build/lib/dryft/signal.py
@@ -1,221 +1,0 @@
-"""
-
-Removes non-linear ground reaction force signal drift in a stepwise manner. It is intended
-for running ground reaction force data commonly analyzed in the field of Biomechanics. The aerial phase before and after
-a given stance phase are used to tare the signal instead of assuming an overall linear trend or signal offset.
-
-Licensed under an MIT License (c) Ryan Alcantara 2019
-
-Distributed here: https://github.com/alcantarar/dryft
-
-
-"""
-import numpy as np
-import matplotlib.pyplot as plt
-import pandas as pd
-
-
-def detrend(force_f, aerial, aerial_loc):
-    """Remove drift from running ground reaction force signal based on aerial phases.
-
-    Parameters
-    ----------
-    force_f : `ndarray`
-        Filtered ground reaction force signal [n,]. Using unfiltered signal may cause unreliable results.
-    aerial : `ndarray`
-        Array of force signal measured at middle of each aerial phase.
-    aerial_loc : `ndarray`
-        Array of frame indexes for values in aerial. output from `aerialforce()`
-
-    Returns
-    -------
-    force_fd : `ndarray`
-        Array with shape of force_f, but with drift removed (detrended).
-
-    Examples
-    --------
-        from dryft import signal
-
-        force_fd = signal.detrend(GRF_filt, aerial_vals, aerial_loc)
-    """
-
-    force_f = force_f.flatten()
-    # Create NaN array with aerial values at respective frame locations
-
-    drift_signal = np.full(force_f.shape, np.nan)
-    drift_signal[aerial_loc] = aerial
-    # Use 3rd order spline to fill NaNs, creating the underlying drift of the signal.
-    drift_signal_p = pd.Series(drift_signal)
-    drift_signal_p = drift_signal_p.interpolate(method = 'spline', order = 3, s = 0, limit_direction= 'both')
-    drift_signal = drift_signal_p.to_numpy()
-    # Subtract drift from force signal
-    force_fd = force_f - drift_signal
-
-    return force_fd
-
-
-def findgoodaerial(stance_begin, stance_end, good_stances):
-    """Locate good aerial phases when bad stance phases are present.
-
-        Parameters
-        ----------
-        stance_begin : `ndarray`
-            Array of frame indexes for start of each stance phase.
-        stance_end : `ndarray`
-            Array of frame indexes for end of each stance phase. Same size as `begin`.
-        good_stances : `ndarray`
-            Boolean array of which stance phases meet min_tc & max_tc requirements.
-
-        Returns
-        -------
-        good_aerial_begin : `ndarray`
-            Array of frame indexes for aerial phase beginnings not connected to bad steps (per min/max_tc requirements).
-        good_aerial_end : `ndarray`
-            Array of frame indexes for aerial phase ends not connected to bad steps (per min/max_tc requirements).
-        """
-    bs = np.where(good_stances == False)
-    aerial_start = np.ones((len(good_stances),), dtype=bool)
-    aerial_end = np.ones((len(good_stances),), dtype=bool)
-
-    aerial_end[bs[0]] = False
-    aerial_end[bs[0] + 1] = False
-
-    aerial_start[bs[0]] = False
-    aerial_start[bs[0] - 1] = False
-
-    good_aerial_begin = stance_end[aerial_start][:-1]
-    good_aerial_end = stance_begin[aerial_end][1:]
-
-    return good_aerial_begin, good_aerial_end
-
-
-def aerialforce(force, stance_begin, stance_end, good_stances):
-    """Calculate force signal at middle of aerial phase of running.
-
-    Parameters
-    ----------
-    force : `ndarray`
-        Filtered vertical ground reaction force (vGRF) signal [n,]. Using unfiltered signal will cause unreliable results.
-    stance_begin : `ndarray`
[…]
