+++ dryft.egg-info/requires.txt
@@ -1,4 +1,0 @@
-numpy
-pandas>=0.24.0
-matplotlib
-scipy
