### 2019-11-27 47d6604
revive link to test notebook. hosted at alcantarar.github.io
