### 2019-11-28 5d6ee14
allow for dependency install via pip/venv, fixes #29, per https://github.com/openjournals/joss-reviews/issues/1910
