### 2019-11-28 5a7752a
add print statement at end of test
