### 2019-11-29 123eab5
discuss dryft compared to other methods more in README, per #34
