### 2019-11-29 990bf99
create simple testscript for MATLAB and incorporate it into sample.m
add simple test script to sample.m which fixes #35 for MATLAB
