### 2019-11-29 3d0b569
Make test run in site-packages
