### 2019-11-29 82c32d1
update readme to separate virtualenv method and pip install method #29
