### 2019-11-29 4e56b86
fixes #24, update pdocs documentation with new test.py
