+++ dryft/sample/__init__.py
@@ -1,0 +1,11 @@
+"""
+
+The `sample` module contains a test script for the `dryft` package.
+
+Licensed under an MIT License (c) Ryan Alcantara 2019
+
+https://alcantarar.github.io
+
+Distributed here: https://github.com/alcantarar/dryft
+
+"""
