### 2019-11-29 #1910 @alcantarar
Thank you @agricolab for your review!

@DARSakthi  and @mwptrsn, commit https://github.com/alcantarar/dryft/commit/4e56b8693a86b371b335bc3d674d8d4fce3f3314 is the latest and greatest.
