### 2019-12-02 #1910 @mwptrsn
> @mwptrsn "Automated Tests" is unchecked in your review checklist; did you have any issues with the testing suite?

I may have missed that/unchecked it by accident. I was able to get the package up and running with no issues.
