### 2019-12-02 #1910 @mwptrsn
The author describes a software package for removing drift in ground force measurements caused by environmental or other factors. The software is well-documented, with both examples and API documentation for both the Python and MATLAB code. The use of a Jupyter Notebook to highlight the use was appreciated. The Python code was easier to set up and get running, but that’s due to the nature of the two environments vs. the software.

I have no issues preventing me from recommending the software and associated paper for publication  in its current state - the previous review was thorough and the concerns have been addressed. However, there are two things that I think could improve the software in the future:

1. The plot of original vs. corrected used in the notebook could be encapsulated in a function. This strikes me as a common analysis that would be performed
2. While an automated smoke test is provided, since the author used synthetic drift in the data used for this experiment, a basic functional test could be created to potentially avoid regressions if the algorithm is changed in the future
