### 2019-12-02 #1910 @DARSakthi
The author presents a novel method for correcting non-linear error in measurements of ground reaction force, and an accompanying software package. The author submits a well written paper reporting necessity, novelty, function, and applicability. I echo the previous reviewers responses, especially in regards to test cases, which I think would be received well if maintained in this repository. Based on the previous edits made and the current state of the paper/repository I am glad to recommend this paper for publication.

For possible future updates, I would recommend including methods to make large data analysis more efficient, beyond Matlab’s automatic parallelism/batch procedures, if the norm for the field involves large data sets or lots of data at once. An additional ReadMe elaborating on the maths behind Spline Interpolation and data subtraction as applied here may be appreciated by some non-specialists.
