### 2019-12-02 #1910 @alcantarar
@mwptrsn and @DARSakthi thanks for the review and suggestions.

@mwptrsn could you elaborate on your second point in an [issue](https://github.com/alcantarar/dryft/issues)?
