### 2019-12-03 #1910 @alcantarar
@Kevin-Mattheus-Moerman
Tagged release v1.1.0 of reviewed software: https://github.com/alcantarar/dryft/releases/tag/v1.1.0
Archived v1.1.0 in Zenodo:  https://zenodo.org/record/3561026
Zenodo DOI: 10.5281/zenodo.3561026
