### 2019-12-03 #1910 @Kevin-Mattheus-Moerman
@alcantarar I have checked the Zenodo archive and its meta-data, and have reread the paper. All looks good so i will now proceed to recommend acceptance in JOSS. I suggest you take this moment to double-check your paper yourself too prior to final acceptance.
