### 2019-12-06 70359fb
final check before JOSS, fixed typos
https://github.com/openjournals/joss-reviews/issues/1910
